+++ R/Aaaa.R
@@ -1,26 +1,26 @@
 #' @importFrom yaml yaml.load_file as.yaml
 #' @importFrom dplyr filter %>% bind_rows
-#' @importFrom tidyselect vars_select eval_select
+#' @importFrom tidyselect vars_select eval_select eval_rename
 #' @importFrom tibble tibble as_tibble
-#' @importFrom dplyr mutate if_else .data desc
+#' @importFrom dplyr mutate if_else .data desc rowwise
 #' @importFrom rmarkdown render pdf_document html_document
 #' @importFrom knitr kable
 #' @importFrom xtable xtable
-#' @importFrom rlang quos set_names exprs as_string expr 
+#' @importFrom rlang quos set_names exprs as_string expr quo_get_expr enquo
+#' @importFrom rlang is_named 
 #' @importFrom assertthat assert_that
 #' @importFrom purrr map map_chr map_df map_if map_lgl
 #' @importFrom purrr imap imap_chr map_int
 #' @importFrom purrr discard compact transpose
 #' @importFrom purrr walk walk2 iwalk keep
-#' @importFrom purrr flatten flatten_chr modify
+#' @importFrom purrr flatten flatten_chr modify imodify
 #' @importFrom glue glue
 #' @importFrom utils type.convert read.csv 
 #' @importFrom utils capture.output head tail
 #' @importFrom tools toTitleCase
 #' @importFrom crayon red green black blue bold italic
 #' @importFrom fs path_rel
-#' @importFrom stringr fixed str_detect str_split_fixed
-#' @importFrom stringr str_count fixed
+#' @importFrom stringr fixed str_detect str_split_fixed str_count fixed 
 #' 
 #' @include utils.R
 NULL
@@ -38,11 +38,21 @@ VALID_SPEC_NAMES <- c(
 
 VALID_SETUP_NAMES <- c(
   "primary_key", "lookup_file", 
-  "description", "sponsor", "projectnumber", 
+  "description", "comment", "sponsor", "projectnumber", 
   "data_path", "data_stem", "name", "spec_file", 
   "spec_path", "glue", "use_internal_db", 
-  "import", "character_last","comment_col"
+  "import", "character_last","comment_col", 
+  "max_nchar_label", "max_nchar_col", "max_nchar_short", 
+  "flags"
 )
+
+ys_control_defaults <- function() {
+  list(
+    max_nchar_label = 40,
+    max_nchar_short = 40,
+    max_nchar_col = 8
+  )
+}
 
 VALID_NS_NAMES <- c(
   "unit", "short", "label", "long", "decode", "comment"
@@ -82,13 +92,15 @@ VALID_NS_NAMES <- c(
 #' 
 #' - `ys.sanitize` a function to use for sanitizing text before processing with 
 #'   latex; see [yspec::pander_table()]
-#' - `ys.col.len` the maximum number of characters in a data frame column 
-#'   name; this defaults to 8 so that columns with 9 or more characters will
-#'   generate an error on load. 
 #' - `ys.fct.suffix` the suffix to add to a column name, used by 
 #'   [yspec::yspec_add_factors()]
 #' - `ys.require.label` if `TRUE`, an error will be generated whenever a column
 #'   is specified without a label
+#'   
+#' **NOTE**: `ys.col.len` was an available in previous versions to set the 
+#' maximum number of characters allowable in a column name. This options has 
+#' been deprecated.  Please use `max_char_col` in `SETUP__:` instead.
+#' 
 #' @docType package
 #' @md
 #' @name yspec