+++ R/load_spec.R
@@ -1,12 +1,32 @@
 
-check_spec_input_col <- function(x, col, env, not_allowed = NULL, ...) {
+get_spec_control <- function(meta = NULL) {
+  ans <- ys_control_defaults()
+  if(is.null(meta)) return(ans)
+  mo <- map(ans, mode)
+  for(nn in names(ans)) {
+    if(.has(nn, meta)) {
+      assert_that(
+        identical(mode(meta[[nn]]), mo[[nn]]),
+        msg = glue("meta field {nn} has wrong type (required: {mo[[nn]]})")
+      )
+      ans[[nn]] <- meta[[nn]]  
+    }
+  }
+  ans
+}
+
+check_spec_input_col <- function(x, col, env, not_allowed = NULL, control, ...) {
   err <- c()
   if(is.null(x)) return()
-  t0 <- nchar(col) <= getOption("ys.col.len",8)
+  t0 <- nchar(col) <= control[["max_nchar_col"]]
   if(!t0) {
     err <- c(
       err,
-      paste0("column name more than ",  getOption("ys.col.len",8), " characters long")
+      paste0(
+        "column name more than ",  
+        control[["max_nchar_col"]], 
+        " characters long"
+      )
     )
   }
   # fields within namespaces have the following pattern: field.namespace
@@ -41,7 +61,7 @@ check_spec_input_col <- function(x, col, env, not_allowed = NULL, ...) {
 
 check_spec_input <- function(x, .fun = check_spec_input_col,
                              context = "spec", ...) {
-  err <- check_for_err(x, .fun)
+  err <- check_for_err(x, .fun, ...)
   if(length(err)==0) return(invisible(NULL))
   file <- basename(get_meta(x)[["spec_file"]])
   file <- paste0("In file: ", file)
@@ -49,7 +69,7 @@ check_spec_input <- function(x, .fun = check_spec_input_col,
 }
 
 
-check_this_col <- function(x,col,env,...) {
+check_this_col <- function(x, col, env, control, ...) {
   err <- c()
   if(.has("values",x)) {
     if(any(x[["values"]] == "<yspec-null>")) {
@@ -80,16 +100,20 @@ check_this_col <- function(x,col,env,...) {
   if(length(x$label) > 1) {
     err <- c(err, "the 'label' field should not be more than length 1")  
   }
-  if(sum(nchar(x$label)) > 40) {
-    err <- c(err, "the 'label' field should not be longer than 40 characters")
+  if(sum(nchar(x$label)) > control[["max_nchar_label"]]) {
+    nmax <- control[["max_nchar_label"]]
+    msg <- "the 'label' field should not be longer than {nmax} characters"
+    err <- c(err, as.character(glue(msg)))
   }
   if(isTRUE(env$require.label)) {
     if(!is.character(x$label)) {
       err <- c(err, "'label' is required for every column, but is missing")  
     }
   }
-  if(sum(nchar(x$short)) > 40) {
-    err <- c(err, "the 'short' field should not be longer than 40 characters")  
+  if(sum(nchar(x$short)) > control[["max_nchar_short"]]) {
+    nmax <- control[["max_char_short"]]
+    msg <- "the 'short' field should not be longer than {nmax} characters"
+    err <- c(err, as.character(glue(msg)))
   }
   if(! all(x[["type"]] %in% c("numeric", "character", "integer"))) {
     bad <- setdiff(x[["type"]],c("numeric", "character", "integer"))
@@ -101,8 +125,8 @@ check_this_col <- function(x,col,env,...) {
   env$err[[col]] <- err
 }
 
-check_spec_cols <- function(x, context = "column") {
-  err <- check_for_err(x, check_this_col)
+check_spec_cols <- function(x, context = "column", ...) {
+  err <- check_for_err(x, check_this_col, ...)
   if(length(err)==0) return(invisible(NULL))
   file <- basename(get_meta(x)[["spec_file"]])
   file <- paste0("In file: ", file)
@@ -111,7 +135,7 @@ check_spec_cols <- function(x, context = "column") {
 
 check_for_err <- function(x, .fun, ...) {
   env <- new.env()
-  env$err <- set_names(vector("list", length(x)),names(x))
+  env$err <- set_names(vector("list", length(x)), names(x))
   env$require.label <- getOption("ys.require.label", FALSE)
   walk2(x, names(x), .fun, env = env, ...)
   err <- discard(as.list(env)$err, is.null)
@@ -148,9 +172,18 @@ capture_file_info <- function(x,file,where = "SETUP__") {
 ##' 
 ##' @md
 ##' @export
-ys_load <- function(file, verbose=FALSE,  ...) {
-  x <- ys_load_file(file, verbose=verbose,...)
-  x <- unpack_spec(x,verbose=verbose)
+ys_load <- function(file, verbose = FALSE, ...) {
+  # not using lifecycle yet; possibly in the future
+  if(!is.null(getOption("ys.col.len", NULL))) {
+    warning(
+      "The option `ys.col.len` has been deprecated; please use the control\n", 
+      "field `max_nchar_col` in SETUP__: instead of the option.", 
+      call. = FALSE
+    )
+  }
+  x <- ys_load_file(file, verbose = verbose,...)
+  x <- unpack_spec(x, verbose = verbose)
+  x <- add_flags(x)
   set_namespace(x, "base")
 }
 
@@ -163,25 +196,27 @@ ys_load_file <- function(file, data_path = NULL, data_stem = NULL, verbose=FALSE
   file <- normalPath(file, mustWork = FALSE)
   if(verbose) verb("~ working on", basename(file))
   x <- try_yaml(file)
-  x <- capture_file_info(x,file)
+  x <- capture_file_info(x, file)
   incoming <- list(...)
   incoming[["data_path"]] <- data_path
   incoming[["data_stem"]] <- data_stem
-  unpack_meta(x,to_update=incoming,verbose=verbose)
+  unpack_meta(x, to_update = incoming, verbose = verbose)
 }
 
 ##' @rdname ys_load
 ##' @export
 load_spec <- function(...) ys_load(...)
 
-unpack_spec <- function(x,verbose=FALSE) {
+unpack_spec <- function(x, verbose = FALSE) {
   
-  check_spec_input(x)
+  control <- pull_meta(x, "control")
+  
+  check_spec_input(x, control = control)
   
   # for looking up column data
   x[] <- imap(x, .f = col_initialize)
   
-  lookup <- ys_get_lookup(x,verbose=verbose)
+  lookup <- ys_get_lookup(x, verbose = verbose)
   
   if(length(lookup) > 0 & verbose) {
     verb(relapse(":",13), relapse(":",30))
@@ -197,27 +232,27 @@ unpack_spec <- function(x,verbose=FALSE) {
   )
   
   x[] <- map(x, unpack_col)
-
+  
   m <- get_meta(x)
   m[["namespace"]] <- list_namespaces(x)
   x <- structure(x, meta = m)
   
-  check_spec_cols(x)
+  check_spec_cols(x, control = control)
   ans <- structure(x, class = "yspec")
   if(.has("import", get_meta(x))) {
-    import <- ys_load(maybe_pull_meta(x,"import"))
-    ans <- c(import,ans,.meta = get_meta(ans))
+    import <- ys_load(maybe_pull_meta(x, "import"))
+    ans <- c(import, ans, .meta = get_meta(ans))
   }
   if(isTRUE(maybe_pull_meta(x, "character_last"))) {
     type <- map_chr(ans, "type")
-    chr <- type=="character"
+    chr <- type == "character"
     comment <- names(ans) %in% maybe_pull_meta(x, "comment_col")
-    ans <- c(ans[(!chr) | comment],ans[chr & (!comment)])
+    ans <- c(ans[(!chr) | comment], ans[chr & (!comment)])
   }
   ans
 }
 
-unpack_meta <- function(x,to_update, verbose=FALSE, ...) {
+unpack_meta <- function(x, to_update, verbose = FALSE, ...) {
   meta <- list()
   metai <- names(x) == "SETUP__"
   if(any(metai)) {
@@ -230,8 +265,8 @@ unpack_meta <- function(x,to_update, verbose=FALSE, ...) {
   }
   if(exists("lookup_file", meta)) {
     assert_that(is.character(meta[["lookup_file"]]))
-    meta[["lookup_file"]] <- file.path(meta[["spec_path"]],meta[["lookup_file"]])
-    meta[["lookup_file"]] <- normalPath(meta[["lookup_file"]],mustWork = FALSE)
+    meta[["lookup_file"]] <- file.path(meta[["spec_path"]], meta[["lookup_file"]])
+    meta[["lookup_file"]] <- normalPath(meta[["lookup_file"]], mustWork = FALSE)
   }
   if(.no("name", meta)) {
     meta[["name"]] <- basename(meta[["spec_file"]])
@@ -261,11 +296,13 @@ unpack_meta <- function(x,to_update, verbose=FALSE, ...) {
     if(.has("project", meta)) verb("  project", meta[["project"]])
     if(.has("sponsor", meta)) verb("  sponsor", meta[["sponsor"]])
   }
-  meta <- update_list(meta,to_update)
+  meta <- update_list(meta, to_update)
   if(exists("import", meta)) {
-    meta[["import"]] <- fs::path_abs(meta[["import"]],meta[["spec_path"]])  
+    meta[["import"]] <- fs::path_abs(meta[["import"]], meta[["spec_path"]])  
   }
   spec_validate_meta(meta)
+  meta[["flags"]] <- validate_flags(meta[["flags"]], valid = names(x))
+  meta[["control"]] <- get_spec_control(meta)
   structure(x, meta = meta)
 }